Add support for reading the DHT22 sensor
Depending on kernel settings the lisp process might get unscheduled
during read. As reading is time critical we will either miss edges,
which leads to the "Timeout waiting for pin low" error, or read
invalid values (due to invalid average calculation) which leads to an
invalid checksum check ("Checksum check failed" error).

Either way, there are restarts that can be invoked to retry reading or
return the last successful read. Note that the sensors is read only
once every three seconds, to prevent any damage. The last succuessful
read is cached and returned, if called with a higher frequency.